Made changes to exclude orgs / service names, calculate average service instance and application instance count.

+++ Usage/src/main/java/io/pivotal/tola/cfapi/Usage/model/AUsage.java
@@ -14,6 +14,6 @@ public class AUsage {
 
     private long totalAis;
     private long totalMbPerAis;
-    private long aiDurationInSecs;
+    private double aiDurationInSecs;
 
 }
