+++ Usage/src/main/java/io/pivotal/tola/cfapi/Usage/model/SpaceUsage.java
@@ -13,6 +13,6 @@ public class SpaceUsage {
     private long totalApps;
     private long totalAis;
     private long totalMbPerAis;
-    private long aiDurationInSecs;
+    private double aiDurationInSecs;
 
 }
